Update README.md
Added a new 'group' of incompatible plugins - this change will become more prominent with the new spigot scoreboard API in 1.13. However it is also a problem with NMS based plugins.
Any plugin that creates team color sidebars will not work - since a player can only be in one team. They have to hook into NTE and get the name of the team the player is in